defconfig: Enable hang detection
Same logic as for enabling the CPU watchdog.

"It may be rare that we get a lockup, but what if it happens in the
middle of the night and your alarm doesn't fire because of it? According
to Sultan, even the tiniest of embedded devices have watchdogs. It fires
rarely enough that it really doesn't matter."

Test: device boots and is stable
Signed-off-by: kdrag0n <dragon@khronodragon.com>