cpu_input_boost: Apply a negative SchedTune boost on display off
Negative SchedTune boosting can reduce power consumption when the
display is off, as performance is not critical in such scenarios. This
can be configured as desired in Kconfig and sysfs. Boost can be as low
as -100, which somewhat resembles the powersave governor according to
LWN: https://lwn.net/Articles/704859/

Test: /dev/stune/schedtune.boost reads 0 when display is on and -50 when
display is off
Test: adjusting suspend_stune_boost parameter works
Signed-off-by: kdrag0n <dragon@khronodragon.com>