drm/msm: klapse: Modify klapse_scaling_rate to use a minute value
Modified to work with mdss port of KLapse (moved changes to mdss port file)

This is more intuitive as compared to a numeric factor.

- Fix an issue with brightness_factor where increasing the value would
dim the colors further until the next pulse. (brightness_factor without
pulse fixes it at the cost of a maximum error of 1 RGB value)
- Bump module version to v4.2

Thanks to @flar2 for the suggestion.

Signed-off-by: kdrag0n <dragon@khronodragon.com>