schedutil: Change rate limits
After some tests in our group we found out that this performs nearly same as
old values. Battery life increased a little bit with this values.
Credits for this to everyone who tested at group and tried his own values.
Hopefully we will get positive feedback at all, if not this will be reverted